Refactor to make the code more testable
