+++ riff-raff/app/notification/FailureNotificationContents.scala
@@ -1,0 +1,60 @@
+package notification
+
+import com.gu.anghammarad.models.{Action, Target}
+import controllers.routes
+import deployment.{NoDeploysFoundForStage, ScheduledDeployError, SkippedDueToPreviousFailure, SkippedDueToPreviousWaitingDeploy}
+import magenta.DeployParameters
+
+import java.util.UUID
+
+object FailureNotificationContents {
+
+  case class NotificationContents(subject: String, message: String, actions: List[Action])
+  case class NotificationContentsWithTargets(notificationContents: NotificationContents, targets: List[Target])
+
+  def problematicDeployUrl(prefix: String, uuid: UUID): String = {
+    val path = routes.DeployController.viewUUID(uuid.toString, verbose = true)
+    prefix + path.url
+  }
+
+  def deployAgainUrl(prefix: String, uuid: UUID): String = {
+    val path = routes.DeployController.deployAgainUuid(uuid.toString)
+    prefix + path.url
+  }
+
+  def scheduledDeployConfigUrl(prefix: String): String = {
+    val path = routes.ScheduleController.list()
+    prefix + path.url
+  }
+
+  def deployFailedNotificationContents(uuid: UUID, parameters: DeployParameters, urlPrefix: String): NotificationContents = {
+    NotificationContents(
+      subject = s"${parameters.deployer.name} failed",
+      message = s"${parameters.deployer.name} for ${parameters.build.projectName} (build ${parameters.build.id}) to stage ${parameters.stage.name} failed.",
+      actions = List(Action("View failed deploy", problematicDeployUrl(urlPrefix, uuid)))
+    )
+  }
+
+  def deployUnstartedNotificationContents(scheduledDeployError: ScheduledDeployError, urlPrefix: String, teamTargetsSearch: (Option[UUID], Option[DeployParameters]) => List[Target], riffRaffTargets: List[Target]): NotificationContentsWithTargets = {
+    val subject = "Scheduled deploy failed to start"
+    def viewProblematicDeploy(uuid: UUID, status: String) = Action(s"View $status deploy", problematicDeployUrl(urlPrefix, uuid))
+    scheduledDeployError match {
+      case SkippedDueToPreviousFailure(record) =>
+        val message = s"A scheduled deploy of ${record.parameters.build.projectName} to ${record.parameters.stage} didn't start because the most recent deploy failed."
+        val redeployAction = Action("Redeploy manually", deployAgainUrl(urlPrefix, record.uuid))
+        val contents = NotificationContents(subject, message, List(viewProblematicDeploy(record.uuid, "failed"), redeployAction))
+        NotificationContentsWithTargets(contents, teamTargetsSearch(Some(record.uuid), Some(record.parameters)))
+      case SkippedDueToPreviousWaitingDeploy(record) =>
+        val message = s"A scheduled deploy of ${record.parameters.build.projectName} to ${record.parameters.stage} failed to start as a previous deploy was still waiting to be deployed."
+        val contents = NotificationContents(subject, message, List(viewProblematicDeploy(record.uuid, "waiting")))
+        NotificationContentsWithTargets(contents, riffRaffTargets)
+      case NoDeploysFoundForStage(projectName, stage) =>
+        val message = s"A scheduled deploy didn't start because RiffRaff has never deployed $projectName to $stage before. " +
+          "Please inform the owner of this schedule as it's likely that they have made a configuration error."
+        val scheduledDeployConfig = Action("View scheduled deploy configuration", scheduledDeployConfigUrl(urlPrefix))
+        val contents = NotificationContents(subject, message, List(scheduledDeployConfig))
+        NotificationContentsWithTargets(contents, riffRaffTargets)
+    }
+  }
+
+}
